requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==0.8.11
requests==2.31.0
openai==1.3.0
langchain==0.0.350
langgraph==0.0.20
python-ollama==0.1.1